Added classes for specialized cargo
