+++ src/main/java/com/alevel/trucking/model/address/Address.java
@@ -1,6 +1,7 @@
 package com.alevel.trucking.model.address;
 
 public class Address {
+    private Long id;
     private City city;
     private Street street;
     private Building building;
