+++ src/main/java/com/alevel/trucking/model/address/Building.java
@@ -1,5 +1,6 @@
 package com.alevel.trucking.model.address;
 
 public class Building {
+    private Long id;
     private int number;
 }
