+++ src/main/java/com/alevel/trucking/model/address/City.java
@@ -1,5 +1,6 @@
 package com.alevel.trucking.model.address;
 
 public class City {
+    private Long id;
     private String name;
 }
