+++ src/main/java/com/alevel/trucking/model/address/Street.java
@@ -1,5 +1,6 @@
 package com.alevel.trucking.model.address;
 
 public class Street {
+    private Long id;
     private String name;
 }
