+++ src/main/java/com/alevel/trucking/model/goods/Goods.java
@@ -1,0 +1,6 @@
+package com.alevel.trucking.model.goods;
+
+public class Goods {
+    private Long id;
+    private int weight;
+}
