+++ src/main/java/com/alevel/trucking/model/goods/GoodsForCoveredTransport.java
@@ -1,0 +1,4 @@
+package com.alevel.trucking.model.goods;
+
+public class GoodsForCoveredTransport extends Goods {
+}
