+++ src/main/java/com/alevel/trucking/model/goods/GoodsForOpenTransport.java
@@ -1,0 +1,4 @@
+package com.alevel.trucking.model.goods;
+
+public class GoodsForOpenTransport extends Goods {
+}
