+++ src/main/java/com/alevel/trucking/model/goods/GoodsForTankTransport.java
@@ -1,0 +1,4 @@
+package com.alevel.trucking.model.goods;
+
+public class GoodsForTankTransport extends Goods {
+}
