--- src/main/java/com/alevel/trucking/model/order/Goods.java
@@ -1,5 +1,0 @@
-package com.alevel.trucking.model.order;
-
-public class Goods {
-    private int weight;
-}
