+++ src/main/java/com/alevel/trucking/model/order/Order.java
@@ -1,13 +1,17 @@
 package com.alevel.trucking.model.order;
 
-import newpost.model.person.Customer;
-import newpost.model.person.Manager;
-import newpost.model.person.driver.Driver;
-import newpost.model.transport.Transport;
+
+import com.alevel.trucking.model.goods.Goods;
+import com.alevel.trucking.model.person.Customer;
+import com.alevel.trucking.model.person.Manager;
+import com.alevel.trucking.model.person.driver.Driver;
+import com.alevel.trucking.model.route.Route;
+import com.alevel.trucking.model.transport.Transport;
 
 import java.util.List;
 
 public class Order {
+    private Long id;
     private Route route;
     private Customer customer;
     private List<Driver> driver;
@@ -15,4 +19,5 @@ public class Order {
     private List<Goods> goods;
     private OrderStatus status;
     private List<Transport> transport;
+    private double cost;
 }
