+++ src/main/java/com/alevel/trucking/model/order/OrderStatus.java
@@ -2,5 +2,5 @@ package com.alevel.trucking.model.order;
 
 public enum OrderStatus {
     IN_PROCESS,
-    DONE;
+    DONE
 }
