--- src/main/java/com/alevel/trucking/model/order/Route.java
@@ -1,10 +1,0 @@
-package com.alevel.trucking.model.order;
-
-import newpost.model.address.Address;
-
-
-public class Route {
-    private Address start;
-    private Address end;
-    private double distance;
-}
