+++ src/main/java/com/alevel/trucking/model/person/driver/DriverLicense.java
@@ -4,6 +4,7 @@ import java.util.Date;
 import java.util.Set;
 
 public class DriverLicense {
+    private Long id;
     private Date dateOfRegistration;
     private Date dateOfFirstRegistration;
     private String typeOfProperty;
