+++ src/main/java/com/alevel/trucking/model/route/Route.java
@@ -1,0 +1,10 @@
+package com.alevel.trucking.model.route;
+
+import com.alevel.trucking.model.address.Address;
+
+public class Route {
+    private Long id;
+    private Address start;
+    private Address end;
+    private double distance;
+}
