+++ src/main/java/com/alevel/trucking/model/transport/Transport.java
@@ -1,6 +1,8 @@
 package com.alevel.trucking.model.transport;
 
 public abstract class Transport {
+    private Long id;
+    private String name;
     private TransportType type;
     private TransportStatus status;
     private String licensePlateNumber;
