+++ src/main/java/com/alevel/trucking/model/transport/TransportStatus.java
@@ -1,4 +1,7 @@
 package com.alevel.trucking.model.transport;
 
 public enum TransportStatus {
+    IN_BOX,
+    IN_ROUTE,
+    IN_REPEIR
 }
