+++ src/main/java/com/alevel/trucking/model/transport/TransportType.java
@@ -1,4 +1,9 @@
 package com.alevel.trucking.model.transport;
 
 public enum TransportType {
+    CONTAINER,
+    COVERED,
+    EVACUATOR,
+    OPEN,
+    TANK
 }
